Edits to the main Class
Added public statements before variables (Its a style thing :))
Removed Extraneous catch statements
Spaced things out, fixed formatting. If you are using eclipse, try Control+F to auto format your files
Changed isPrime test methods to make a little more sense, variable is too vague.
Made your fileWriter static and put it with the fields instead of as a local variable in the method. It just seems right for it to be in the fields instead!